oops

+++ admin.py
@@ -11,6 +11,8 @@ class Request:
         self._request_timestamp = row.request_timestamp
         self._club_id = row.club_id
         self._user_id = row.user_id
+        self._name = row.name
+        self._year = row.year
 
     def get_request_timestamp(self):
         return self._request_timestamp
@@ -20,6 +22,12 @@ class Request:
     
     def get_user_id(self):
         return self._user_id
+
+    def get_name(self):
+        return self._name
+    
+    def get_year(self):
+        return self._year
 
 
 #-----------------------------------------------------------------------
@@ -33,7 +41,8 @@ if DATABASE_URL and DATABASE_URL.startswith("postgres://"):
 # and then figure out the button stuff so they are actually clickable
 # fix navbar information
 def delete_request(user_id, club_id):
-
+    if user_id is None:
+        return
     engine = sqlalchemy.create_engine(DATABASE_URL)
     try:
         with sqlalchemy.orm.Session(engine) as session:
@@ -45,22 +54,24 @@ def delete_request(user_id, club_id):
         engine.dispose()
 
 def approve_request(user_id, club_id):
-    delete_request(user_id, club_id)
-    engine = sqlalchemy.create_engine(DATABASE_URL)
+    if user_id is None:
+        return
 
+    engine = sqlalchemy.create_engine(DATABASE_URL)
     try:
         with sqlalchemy.orm.Session(engine) as session:
-            validatedUser = database.Users_Clubs(username = "yparikh",
+            validatedUser = database.Users_Clubs(username = user_id,
                         club_id = club_id)
             session.add(validatedUser)
             session.commit()
             print("added to database")
+        
+        delete_request(user_id, club_id)
     finally:
         engine.dispose()
 
 def get_requests():
     engine = sqlalchemy.create_engine(DATABASE_URL)
-
     try:
         with sqlalchemy.orm.Session(engine) as session:
             query = session.query(database.Requests)
@@ -77,13 +88,13 @@ def create_request(user_id, club_id):
     engine = sqlalchemy.create_engine(DATABASE_URL)
     try:
         with sqlalchemy.orm.Session(engine) as session:
-            query = session.query(database.Requests).where((database.Requests.user_id == user_id)&(database.Requests.club_id == club_id))
-
             if check_request(user_id, club_id):
                 print("request already exists")
                 return 
             else:
+                new_user = database.User(user_id = user_id)
                 new_request = database.Requests(request_timestamp = datetime.now(), user_id = user_id, club_id = club_id)
+                session.add(new_user)
                 session.add(new_request)
                 session.commit()
                 print("Request added")
